+++ cypress/integration/lotto/lotto.spec.js
@@ -1,14 +1,8 @@
-// import {
-//   DUPLICATE_WINNING_NUMBER,
-//   HAS_A_WHITESPACE_MESSAGE,
-//   LESS_THAN_TICKET_PRICE_MESSAGE,
-//   INPUT_NOT_COMPLETED,
-//   EXCEED_RANGE_NUMBER,
-// } from '../../../src/js/lib/constants/alertMessage';
-// import {
-//   TICKET_MIN_NUMBER,
-//   TICKET_MAX_NUMBER,
-// } from '../../../src/js/lib/constants/lotto';
+import {
+  DUPLICATE_WINNING_NUMBER,
+  EXCEED_RANGE_NUMBER,
+  LESS_THAN_TICKET_PRICE_MESSAGE,
+} from '../../../src/js/lib/constants/alertMessage';
 
 context('Actions', () => {
   beforeEach(() => {
@@ -21,17 +15,11 @@ context('Actions', () => {
     cy.get('#ticket-list > div').should('have.length', 5);
   });
 
-  it('공백은 입력할 수 없다.', () => {
-    cy.window().then(window => cy.stub(window, 'alert').as('alert'));
-    cy.get('button[name=payment-button]').click();
-    cy.get('@alert').should('be.calledWith', 'HAS_A_WHITESPACE_MESSAGE');
-  });
-
-  it('0과 음수는 입력할 수 없다.', () => {
+  it('1000원 미만의 금액은 입력할 수 없다.', () => {
     cy.window().then(window => cy.stub(window, 'alert').as('alert'));
     cy.get('input[name=payment-input]').type('0');
     cy.get('button[name=payment-button]').click();
-    cy.get('@alert').should('be.calledWith', 'LESS_THAN_TICKET_PRICE_MESSAGE');
+    cy.get('@alert').should('be.calledWith', LESS_THAN_TICKET_PRICE_MESSAGE);
   });
 
   it('한 번 로또를 구입하면 다시 구매할 수 없다.', () => {
@@ -88,7 +76,7 @@ context('Actions', () => {
   it('당첨번호는 1~45 사이의 숫자여야한다.', () => {
     cy.window().then(window => cy.stub(window, 'alert').as('alert'));
     cy.get('.winning-number[name=first]').type(99);
-    cy.get('@alert').should('be.calledWith', 'EXCEED_RANGE_NUMBER');
+    cy.get('@alert').should('be.calledWith', EXCEED_RANGE_NUMBER);
   });
 
   it('당첨번호는 중복될 수 없다.', () => {
@@ -100,12 +88,6 @@ context('Actions', () => {
     });
     cy.get('.bonus-number').type(34);
     cy.get('#result').click();
-    cy.get('@alert').should('be.calledWith', 'DUPLICATE_WINNING_NUMBER');
-  });
-
-  it('결과확인 전에 구입 금액과 당첨번호를 모두 입력해야 한다.', () => {
-    cy.window().then(window => cy.stub(window, 'alert').as('alert'));
-    cy.get('#result').click();
-    cy.get('@alert').should('be.calledWith', 'INPUT_NOT_COMPLETED');
+    cy.get('@alert').should('be.calledWith', DUPLICATE_WINNING_NUMBER);
   });
 });
