+++ index.html
@@ -100,6 +100,7 @@
             </div>
             <button
               type="submit"
+              id="result"
               class="open-result-modal-button mt-5 btn btn-cyan w-100"
             >
               결과 확인하기
