+++ src/js/handlers/lottoPurchase.js
@@ -1,57 +1,33 @@
-import { lottoData } from '../model/lottoData.js';
+import { lotto } from '../model/lotto.js';
 import $ from '../lib/utils/dom.js';
 import { LESS_THAN_TICKET_PRICE_MESSAGE } from '../lib/constants/alertMessage.js';
-import {
-  TICKET_PRICE,
-  TICKET_NUMBER_AMOUNT,
-  TICKET_MIN_NUMBER,
-  TICKET_MAX_NUMBER,
-} from '../lib/constants/lotto.js';
+import { TICKET_PRICE } from '../lib/constants/lotto.js';
+import { getTicketNumber } from '../lib/utils/lotto.js';
+import { createTicketTemplate } from '../lib/utils/template.js';
+
+const updateTicketListView = tickets => {
+  const ticketsTemplate = tickets.reduce(
+    (acc, ticket) => acc + createTicketTemplate(ticket),
+    ''
+  );
+
+  $('#ticket-list').innerHTML = ticketsTemplate;
+  $('#ticket-count').innerHTML = `총 ${tickets.length}개를 구매하였습니다.`;
+  $('#toggle-detail-mode').classList.remove('hide');
+};
 
 const lottoPurchaseHandler = event => {
   event.preventDefault();
   const paymentInput = event.target.elements['payment-input'].value;
+  const ticketAmount = Math.floor(Number(paymentInput) / TICKET_PRICE);
 
   if (Number(paymentInput) < TICKET_PRICE) {
     alert(LESS_THAN_TICKET_PRICE_MESSAGE);
     return;
   }
 
-  lottoData.ticketAmount = Math.floor(Number(paymentInput) / TICKET_PRICE);
-  lottoData.tickets = [...Array(lottoData.ticketAmount)].map(() =>
-    getTicketNumber()
-  );
-
-  const ticketsHTML = lottoData.tickets
-    .map(ticketNumber => createTicktHTML(ticketNumber))
-    .join('');
-
-  $('#ticket-list').innerHTML = ticketsHTML;
-  $(
-    '#ticket-count'
-  ).innerHTML = `총 ${lottoData.ticketAmount}개를 구매하였습니다.`;
-  $('#toggle-detail-mode').classList.remove('hide');
+  lotto.setTickets([...Array(ticketAmount)].map(() => getTicketNumber()));
+  updateTicketListView(lotto.tickets);
 };
-
-function createTicktHTML(ticketNumber) {
-  return `<div>
-            <span class="mx-1 text-4xl">🎟️ </span>
-            <span class="ticket-number hide">${ticketNumber.join(', ')}</span>
-          </div>`;
-}
-
-function getTicketNumber() {
-  const ticketNumber = new Set();
-
-  while (ticketNumber.size < TICKET_NUMBER_AMOUNT) {
-    ticketNumber.add(getRandomNumber(TICKET_MIN_NUMBER, TICKET_MAX_NUMBER));
-  }
-
-  return [...ticketNumber];
-}
-
-function getRandomNumber(min, max) {
-  return Math.floor(Math.random() * (max - min + 1) + min);
-}
 
 export default lottoPurchaseHandler;
