+++ src/js/lib/utils/dom.js
@@ -1,3 +1,7 @@
-const $ = selector => document.querySelector(selector);
+const $ = selector => {
+  const elements = document.querySelectorAll(selector);
+
+  return elements.length === 1 ? elements[0] : elements;
+};
 
 export default $;
