+++ src/js/lib/utils/template.js
@@ -1,0 +1,6 @@
+export function createTicketTemplate(ticketNumber) {
+  return `<div>
+            <span class="mx-1 text-4xl">🎟️ </span>
+            <span class="ticket-number hide">${ticketNumber.join(', ')}</span>
+          </div>`;
+}
