+++ src/js/lib/utils/validation.js
@@ -1,0 +1,8 @@
+function hasDuplicate(array) {
+  const arrayLength = array.length;
+  const setLength = new Set(array).size;
+
+  return setLength < arrayLength;
+}
+
+export { hasDuplicate };
