+++ src/js/model/lotto.js
@@ -1,0 +1,6 @@
+export const lotto = {
+  tickets: [],
+  setTickets: newTickets => {
+    this.tickets = newTickets;
+  },
+};
