--- src/js/model/lottoData.js
@@ -1,4 +1,0 @@
-export const lottoData = {
-  tickets: [],
-  ticketAmount: 0,
-};
